fix: modify page, global css for token generation page
