+++ FrontEnd/app/styles/layout.module.css
@@ -7,7 +7,7 @@
     'footer';
   grid-template-rows: auto auto 1fr 36px;
   align-items: center;
-  min-height: 100vh;
+  justify-items: center;
 }
 
 .logo {
@@ -36,42 +36,19 @@
   justify-self: center;
 }
 
-.nav {
-  grid-area: nav;
+.virtualKeyboard {
   display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  row-gap: 1rem;
+  column-gap: 1rem;
+  width: 24rem;
   align-items: center;
   justify-content: center;
-  gap: 8px;
-  padding: 8px;
-  font-size: calc(10px + 2vmin);
 }
 
-.link:hover {
-  text-decoration: underline;
-}
-
-.link {
-  color: #704cb6;
-}
-
-.link.active {
-  text-decoration: underline;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .logo {
-    animation: logo-float infinite 3s ease-in-out;
-  }
-}
-
-@keyframes logo-float {
-  0% {
-    transform: translateY(0);
-  }
-  50% {
-    transform: translateY(10px);
-  }
-  100% {
-    transform: translateY(0px);
-  }
+.buttonLayout {
+  display: flex;
+  flex-direction: row;
+  row-gap: 0.5rem;
 }
